app/requirements.txt
autopep8==1.6.0
gspread==5.3.2
gspread-dataframe
pandas==1.4.2
PyMySQL==1.0.2
SQLAlchemy==1.4.36
python-dotenv